Lade till två DTOs. Se beskrivning för förklaring
What is a DTO?

In short, a simple class that's used to pass data between layers. Maybe your domain models have a large set of properties, but you only need to display five of them to the end user. Then you map this domain object to a Dto and return that to the UI.

You also never want to expose domain models directly to the end user. Mainly because changing the domain would then go through all the way to the end user directly, which could be really bad.

https://www.reddit.com/r/dotnet/comments/nfk3ry/what_is_actually_a_dto_and_when_should_i_use_it/